exclude lib from Magento checkstyle
Until such time that composer use becomes the standard for installation for Magento extensions the obvious place to keep library code is under {MAGENTO_ROOT}/lib. Currently this is subjected to the same Magento style checks. 

In my view this is less than ideal. If I want to score well I would have to do one of the following:

1.) fork the original project and adjust to Magento styling = a fork just to maintain styling is not efficient - also see 3.)
2.) attempt to enforce Magento styling on the original project = why should the original project that works across a range of frameworks adopt this? they are free to use whatever style they like and can't cater to all downstream projects simultaneously and there is no style fits all for the php world
3.) don't use the project and re-invent the wheel with a Magento specific solution = DRY is powerful and the php world is only slowly reaping the benefits of this, this would be a step backwards and likely result in a subpar solution (smaller audience, less widely tested, less history = more bugs still to be ironed out)
4.) make Judge exclude lib from Magento checkstyle checks

Please note that I agree with subjecting lib code to the same checks of the other plugins (for unit testing there needs to be some further thought as well since using EcomDev_PhpUnit does not make sense for a non-Magento project) style is simply the most diverging (Tabs vs Spaces and line length for example).